requirements.txt
librosa==0.11.0
PyQt5_sip==12.17.0
waapi_client==0.6
